Remove useLastMove hook
After some careful thought, I decided that this hook was simply too
complicated to implement as a diff between two FENs. In order to do it
properly, a chess board would need to keep track of the moves a player
made _as they were made_. That seems beyond the scope of this library.